1.5.53 release (only C64.emu has changed)
* C64.emu: Use -fno-finite-math-only for compiling resid/dac.cc to avoid miscompile when building with -Ofast
* C64.emu: Fix possible signed integer overflow in Filter::solve_gain()

+++ C64.emu/src/vice/resid/filter.h
@@ -1370,7 +1370,7 @@ int Filter::solve_gain(opamp_t* opamp, int n, int vi, int& x, model_filter_t& mf
     // The dividend is scaled by m^2*2^27.
     int f = a*int(unsigned(b_vx)*unsigned(b_vx) >> 12) - c - int(unsigned(b_vo)*unsigned(b_vo) >> 5);
     // The divisor is scaled by m*2^11.
-    int df = (b_vo*(dvx + (1 << 11)) - a*(b_vx*dvx >> 7)) >> 15;
+    int df = (unsigned(b_vo)*(dvx + (1 << 11)) - unsigned(a)*(b_vx*dvx >> 7)) >> 15;
     // The resulting quotient is thus scaled by m*2^16.
 
     // Newton-Raphson step: xk1 = xk - f(xk)/f'(xk)
